Refactor routing logic from Main to App component + delete useState in CustomTimer
