+++ медиа/form.html
@@ -1,0 +1,130 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE-edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Рок-группа Asking Alexandria</title>
+</head>
+<body>
+    <h1>Asking Alexandria</h1>
+    <p>
+        Asking Alexandria — британская рок-группа, собранная в 2008 году в городе Йорк, Великобритания. 
+        С момента начала сотрудничества с лейблом Sumerian Records в 2009 году переехала в США, 
+        выпустила 7 студийных альбомов, последний из которых увидел свет 23 августа 2023 года.
+    </p>
+
+    <audio src="./Asking Alexandria-Here I Am-kissvk.com.mp3" controls></audio>
+
+    <h3>Состав группы</h3>
+
+    <img src="./hg2inh2ssdl1kf2x5nqfamctoymkhsgq.webp" width="700" height="500" alt="Группа Asking Alexandria">
+
+    <ul>Нынешний состав
+        <li>
+            Бен Брюс — соло-гитара, дополнительный вокал (2008—настоящее время), клавишные, программирование 
+            (2009—2011)
+        </li>
+        <li>
+            Дэнни Уорсноп — вокал (2008—2015, 2016—настоящее время); дополнительная гитара 
+            (2013—2015, 2016—настоящее время); клавишные, программирование (2009—2011)
+        </li>
+        <li>
+            Джеймс Касселлс — ударные (2008—настоящее время)
+        </li>
+        <li>
+            Кэмерон Лидделл — ритм-гитара (2008—настоящее время)
+        </li>
+        <li>
+            Сэм Бэттли — бас-гитара (2009—настоящее время)
+        </li>
+    </ul>
+
+    <ol>Бывшие участники
+        <li>Райан Биннс — клавишные (2008—2009)</li>
+        <li>Джо Ланкастер — бас-гитара (2008—2009)</li>
+        <li>Денис «Stoff» Шафоростов — вокал (2015—2016)</li>
+    </ol>
+
+    <h3>Дискография</h3>
+    <ul>Студийные альбомы
+        <li>
+            <a 
+                href="https://ru.wikipedia.org/wiki/Stand_Up_and_Scream" 
+                target="_blank" 
+                title="Переход на сайт"
+            >
+                Stand Up and Scream
+            </a>
+        </li>
+        <li>
+            <a 
+                href="https://ru.wikipedia.org/wiki/Reckless_%26_Relentless" 
+                target="_blank" 
+                title="Переход на сайт"
+            >
+                Reckless & Relentless
+            </a>
+        </li>
+        <li>
+            <a 
+                href="https://ru.wikipedia.org/wiki/From_Death_to_Destiny" 
+                target="_blank" 
+                title="Переход на сайт"
+            >
+                From Death to Destiny
+            </a>
+        </li>
+        <li>
+            <a 
+                href="https://ru.wikipedia.org/wiki/The_Black" 
+                target="_blank" 
+                title="Переход на сайт"
+            >
+                The Black
+            </a> 
+        </li>
+        <li>
+            <a 
+                href="https://ru.wikipedia.org/wiki/Asking_Alexandria_(%D0%B0%D0%BB%D1%8C%D0%B1%D0%BE%D0%BC)" 
+                target="_blank" 
+                title="Переход на сайт"
+            > 
+                Asking Alexandria
+            </a>
+        </li>
+        <li>
+            <a 
+                href="https://en.wikipedia.org/wiki/Like_a_House_on_Fire" 
+                target="_blank" 
+                title="Переход на сайт"
+            >
+                Like a House on Fire
+            </a>
+        </li>
+        <li>
+            <a 
+                href="https://ru.wikipedia.org/wiki/See_What%E2%80%99s_on_the_Inside" 
+                target="_blank" 
+                title="Переход на сайт"
+            >
+                See What is on the Inside
+            </a>
+        </li>
+        <li>
+            <a 
+                href="https://en.wikipedia.org/wiki/Where_Do_We_Go_from_Here%3F_(Asking_Alexandria_album)" 
+                target="_blank" 
+                title="Переход на сайт"
+            >
+                Where Do We Go from Here?
+            </a>
+        </li>
+    </ul>
+    <audio src="./Asking Alexandria-I Won-t Give In-kissvk.com.mp3" controls></audio>
+    <video 
+        src="./ASKING ALEXANDRIA - Alone In A Room (Official Music Video).mp4" 
+        controls poster="./asking-alexandria_9_orig.webp">
+    </video>
+</body>
+</html>
